Refactor history and xgene controllers and services
This commit refactors the history and xgene controllers and services to add new endpoints for retrieving all histories and all subjects respectively. The HistoryController now includes a new GET endpoint "all" which returns a list of all history records. Similarly, the XgeneController includes a new GET endpoint "all" which returns a list of all subjects. The corresponding changes have been made in the HistoryService and XgeneService classes to implement the new functionality.